rootReducer & store 분리 / App에 persist store 적용
