+++ packages/client/src/modules/index.ts
@@ -1,39 +1,5 @@
-import {
-  applyMiddleware,
-  combineReducers,
-  createStore,
-} from 'redux'
-import friends from 'modules/friends'
-import profile from 'modules/profile'
-import login from 'modules/login'
-import room from 'modules/room'
-import chat from 'modules/chat'
+import rootReducer from 'modules/root-reducer'
 
-import { composeWithDevTools } from 'redux-devtools-extension'
-import logger from 'redux-logger'
-import createSagaMiddleware from 'redux-saga'
-import rootSaga from 'modules/sagas'
-
-const rootReducer = combineReducers({
-  friends,
-  profile,
-  login,
-  room,
-  chat,
-})
-const sagaMiddleware = createSagaMiddleware()
-
-const middlewares = []
-middlewares.push(sagaMiddleware)
-
-if (process.env.NODE_ENV !== 'production') {
-  middlewares.push(logger)
-}
-const store = createStore(
-  rootReducer,
-  composeWithDevTools(applyMiddleware(...middlewares)),
-)
-sagaMiddleware.run(rootSaga)
-export default rootReducer
 export type RootState = ReturnType<typeof rootReducer>;
-export { store }
+export { default as Store } from 'modules/store'
+export { default as rootReducer } from 'modules/root-reducer'
